Code to demonstrate a bug in the parse node generation
Basically, the problem is as follows: in a sequence(), we get to generate nodes
even if the sequence as a whole fails.

We need to be able to detect, in the builder, that we are in a sequence, and if
this is the case, scrap all children builders which could have been generated
until that point.

Using the stack here is not a problem, since SequenceMatcher would just reset
the stack to what it was before if one rule failed. But not here.