Fix issue #7
Instead of building parse nodes only when we see a @GenerateNode annotation, we
build a context each and every time. Act accordingly depending on whether a
@GenerateNode annotation exists for the current context.

Take the opportunity to greatly simplify the tree building: we now only need a
stack of contexts, we don't need to retain the levels anymore.